backend/requirements.txt
# Web 框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Docker 客户端
docker>=6.1.0

# 异步支持
aiofiles>=23.0.0

# AI/ML
openai>=1.0.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# 工具
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0